+++ src/test/resources/sample/TestPlugin.java
@@ -1,0 +1,12 @@
+package ru.vyarus;
+
+import org.gradle.api.Plugin;
+import org.gradle.api.Project;
+
+public class TestPlugin implements Plugin<Project> {
+
+    @Override
+    public void apply(Project target) {
+
+    }
+}