+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Set UTF-8 encoding for JavaCompile and GroovyCompile tasks
 * Set file.encoding=UTF-8 system property for Test tasks
 * Set UTF-8 encoding for javadoc task (encoding, charSet, docencoding) 
+* Gradle plugin projects compatibility fixes:
+    - When used with java-gardle-plugin, re-use pluginMaven publication instead of creating
+        new one (because java-gardle-plugin hardcode publication name and it has to init it because of alias publications )
+    - Plugin-publish will not create his own javadoc and sources tasks (so java-lib tasks will be used)
+    - Overall, single maven publication used and only one source and javadoc (groovydoc) tasks.
 
 ### 1.1.2 (2018-07-22)
 * Fix missed pom dependencies